move label package from internal to pkg
